Create a dev Canvas API access token.
If you want to setup another dev app to be able to call into the Canvas API
for this app in dev, use the access token added in this commit.

E.g. in platform, you can configure the following for it to be able to
this this API:
CANVAS_TOKEN=XAlbyObifoe76wJECtpLDGEvIVViPVklRnhAkWvUFIm8957NSS5eonRn5oYGqb0y